+++ CHANGELOG.md
@@ -5,11 +5,12 @@ All notable changes to this project are documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## Unreleased
 
 ### Added
 - `DuplicatedImport` detector: PR [#381](https://github.com/nowarp/misti/pull/381)
 - `TransitiveImport` detector: Issue [#382](https://github.com/nowarp/misti/issues/382)
+- Support for external tools: Users can now load custom tools using the `-t` option with a `path:className` format, similar to how external detectors work: Issue [#215](https://github.com/nowarp/misti/issues/215)
 - `SuboptimalSend`: Suggest replacing `self.forward` with `self.{notify,reply}`: Issue [#386](https://github.com/nowarp/misti/issues/386)
 
 ### Changed
@@ -20,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - `ImportGraph`: Set correct file paths of import nodes for `.func` files: PR [#381](https://github.com/nowarp/misti/pull/381)
 - `EtaLikeSimplifications`: Confusing help string for Boolean literals: Issue [#392](https://github.com/nowarp/misti/issues/392)
+
 ## [0.8.3] - 2025-04-26
 
 ### Added